+++ packages/loading-skeleton/docs/LoadingHeading.stories.tsx
@@ -1,14 +1,12 @@
 import React from "react"
-import { ComponentStory, Story } from "@storybook/react"
-import { withDesign } from "storybook-addon-designs"
+import { StoryFn } from "@storybook/react"
 import { Heading } from "@kaizen/typography"
 import { StoryWrapper } from "../../../storybook/components/StoryWrapper"
-import { CATEGORIES, SUB_CATEGORIES } from "../../../storybook/constants"
-import { figmaEmbed } from "../../../storybook/helpers"
 import { LoadingHeading } from ".."
 
 export default {
-  title: `${CATEGORIES.components}/${SUB_CATEGORIES.loadingSkeleton}/Loading Headings`,
+  tags: ["autodocs"],
+  title: "Components/Loading/Loading Headings",
   component: LoadingHeading,
   parameters: {
     docs: {
@@ -16,20 +14,16 @@ export default {
         component: 'import { LoadingHeading } from "@kaizen/loading-skeleton"',
       },
     },
-    ...figmaEmbed(
-      "https://www.figma.com/file/eZKEE5kXbEMY3lx84oz8iN/%E2%9D%A4%EF%B8%8F-UI-Kit%3A-Heart?node-id=4496%3A1"
-    ),
   },
-  decorators: [withDesign],
 }
 
-export const DefaultLoadingHeading: ComponentStory<
-  typeof LoadingHeading
-> = args => <LoadingHeading {...args} />
+export const DefaultLoadingHeading: StoryFn<typeof LoadingHeading> = args => (
+  <LoadingHeading {...args} />
+)
 DefaultLoadingHeading.storyName = "Loading Heading"
 DefaultLoadingHeading.args = { variant: "heading-1" }
 
-const StickerSheetTemplate: Story<{ isReversed: boolean }> = ({
+const StickerSheetTemplate: StoryFn<{ isReversed: boolean }> = ({
   isReversed,
 }) => (
   <StoryWrapper isReversed={isReversed}>