+++ packages/notification/src/GlobalNotification.tsx
@@ -1,9 +1,9 @@
-import * as React from "react"
+import React from "react"
 import GenericNotification, {
   NotificationType,
 } from "./components/GenericNotification"
 
-type GlobalNotificationProps = {
+export type GlobalNotificationProps = {
   type: NotificationType
   children: React.ReactNode
   onHide?: () => void
@@ -11,11 +11,11 @@ type GlobalNotificationProps = {
   persistent?: boolean
 }
 
-const GlobalNotification = (props: GlobalNotificationProps): JSX.Element => (
-  <GenericNotification style="global" {...props} />
-)
+export const GlobalNotification = (
+  props: GlobalNotificationProps
+): JSX.Element => <GenericNotification style="global" {...props} />
 GlobalNotification.defaultProps = {
   persistent: false,
 }
 
-export default GlobalNotification
+GlobalNotification.displayName = "GlobalNotification"