+++ packages/notification/src/index.ts
@@ -1,6 +1,6 @@
-export { default as GlobalNotification } from "./GlobalNotification"
-export { default as InlineNotification } from "./InlineNotification"
-export { default as ToastNotification } from "./ToastNotification"
+export * from "./GlobalNotification"
+export * from "./InlineNotification"
+export * from "./ToastNotification"
 export {
   addToastNotification,
   removeToastNotification,