+++ storybook/components/DocsContainer/ComponentDocsTemplate/subcomponents/Title/Title.tsx
@@ -1,0 +1,21 @@
+import React from "react"
+
+export type TitleProps = {
+  //   context: DocsContainerProps["context"]
+  // TODO: Find correct type
+  context: any
+}
+
+export const Title = ({ context }: TitleProps): JSX.Element => {
+  const titleHierarchy = context.attachedCSFFile.meta.title
+  const title = titleHierarchy.split("/")
+
+  return (
+    <>
+      <h1>{title[title.length - 1]}</h1>
+      <br />
+    </>
+  )
+}
+
+Title.displayName = "Title"