+++ draft-packages/empty-state/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [5.3.36](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-empty-state@5.3.35...@kaizen/draft-empty-state@5.3.36) (2023-04-06)
+
+**Note:** Version bump only for package @kaizen/draft-empty-state
+
+
+
+
+
 ## [5.3.35](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-empty-state@5.3.34...@kaizen/draft-empty-state@5.3.35) (2023-04-05)
 
 **Note:** Version bump only for package @kaizen/draft-empty-state