+++ draft-packages/tag/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [3.4.10](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-tag@3.4.9...@kaizen/draft-tag@3.4.10) (2023-04-06)
+
+**Note:** Version bump only for package @kaizen/draft-tag
+
+
+
+
+
 ## [3.4.9](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-tag@3.4.8...@kaizen/draft-tag@3.4.9) (2023-04-05)
 
 **Note:** Version bump only for package @kaizen/draft-tag